+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Issue 53: FileLoader supports multipart/byteranges
 * Issue 54: Support If-Match, If-None-Match, If-Modified-Since and If-Unmodified-Since
 * Issue 55: Support If-Range
+* Issue 56: Generalize FileLoader to support the loading from different resources, such as database
 
 **0.8.4**